requirements.txt
Flask
joblib
pandas
numpy
scikit-learn==1.3.2
requests
google-generativeai